refactor: migrate qre research read-only boundary
