+++ packages/qre_research/__init__.py
@@ -1,0 +1,4 @@
+"""QRE research package read-only boundary seeds."""
+
+from __future__ import annotations
+
