+++ tests/architecture/test_package_migration_001_target_layout.py
@@ -45,7 +45,6 @@ README_REQUIRED_SECTIONS = (
 )
 
 SCAFFOLD_ONLY_TARGETS = (
-    REPO_ROOT / "packages" / "qre_research",
     REPO_ROOT / "packages" / "qre_execution_sim",
     REPO_ROOT / "packages" / "qre_shadow",
     REPO_ROOT / "packages" / "qre_paper",
@@ -139,10 +138,22 @@ def test_package_migration_001_qre_data_has_only_bounded_read_only_seed() -> Non
     assert files == ["README.md", "__init__.py", "contracts.py"], target
 
 
+def test_package_migration_001_qre_research_has_only_bounded_read_only_seed() -> None:
+    target = REPO_ROOT / "packages" / "qre_research"
+    files = sorted(
+        path.relative_to(target).as_posix()
+        for path in target.rglob("*")
+        if path.is_file() and "__pycache__" not in path.parts
+    )
+
+    assert files == ["README.md", "__init__.py", "universe.py"], target
+
+
 def test_package_migration_001_scanner_classifies_target_paths() -> None:
     assert classify_path("apps/control-plane/README.md") == DOMAIN_CONTROL_PLANE
     assert classify_path("packages/ade_governance/README.md") == DOMAIN_ADE
     assert classify_path("packages/qre_research/README.md") == DOMAIN_QRE
+    assert classify_path("packages/qre_research/universe.py") == DOMAIN_QRE
     assert classify_path("packages/qre_data/README.md") == DOMAIN_QRE
     assert classify_path("packages/qre_data/contracts.py") == DOMAIN_QRE
     assert classify_path("packages/qre_artifacts/README.md") == DOMAIN_QRE
@@ -155,7 +166,7 @@ def test_package_migration_001_scanner_classifies_target_paths() -> None:
     assert classify_path("packages/qre_paper/README.md") == DOMAIN_EXECUTION
     assert classify_path("packages/qre_live/README.md") == DOMAIN_EXECUTION
 
-    assert classify_module("packages.qre_research.contracts") == DOMAIN_QRE
+    assert classify_module("packages.qre_research.universe") == DOMAIN_QRE
     assert classify_module("packages.qre_data.contracts") == DOMAIN_QRE
     assert classify_module("packages.qre_artifacts.public_outputs") == DOMAIN_QRE
     assert classify_module("packages.qre_policy.authority_views") == DOMAIN_QRE
